feat(ReservationTime) : 권한에 따른 시간 추가/삭제 인가 구현

+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,6 +1,5 @@
 package roomescape.controller;
 
-import jakarta.websocket.server.PathParam;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.HttpStatus;
@@ -14,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.reservationTime.ReservationTimeRequesetDto;
+import roomescape.dto.reservationTime.ReservationTimeRequestDto;
 import roomescape.dto.reservationTime.ReservationTimeResponseDto;
-import roomescape.repository.reservationTime.ReservationTimeRepository;
+import roomescape.exception.ForbiddenAccessException;
 import roomescape.service.ReservationService;
 
 @RestController
@@ -38,19 +37,32 @@ public class ReservationTimeController {
     }
 
     @PostMapping
-    @ResponseStatus(HttpStatus.OK)
-    public ReservationTimeResponseDto addReservationTime(@RequestBody ReservationTimeRequesetDto requestDto) {
+    @ResponseStatus(HttpStatus.CREATED)
+    public ReservationTimeResponseDto addReservationTime(
+            @RequestBody ReservationTimeRequestDto requestDto,
+            @RequestParam("role") String role
+    ) {
+        if (!role.equals("admin")) {
+            throw new ForbiddenAccessException("테마 추가는 관리자만 가능합니다.");
+        }
+
         ReservationTime time = reservationService.addReservationTime(requestDto);
         return ReservationTimeResponseDto.from(time);
     }
 
     @DeleteMapping("{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteReservationTime(@PathVariable Long id) {
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteReservationTime(
+            @PathVariable Long id,
+            @RequestParam("role") String role
+    ) {
+        if (!role.equals("admin")) {
+            throw new ForbiddenAccessException("테마 삭제는 관리자만 가능합니다.");
+        }
+
         reservationService.deleteReservationTime(id);
     }
 
-    // /times/available?date=2026-05-08&themeId=1
     @GetMapping("available")
     @ResponseStatus(HttpStatus.OK)
     public List<ReservationTimeResponseDto> getAvailableTimes(
